Support building channel when multiple collections contain term
I wanted to use this to build a channel for a "smart collection" that contained the term "Animation"; however it failed because .search() includes results when it isn't an exact match, so I also got results back for "Sony Animation", "Disney Animation", etc. collections, which caused the script to fail. These edits fix this, and require an exact match for the supplied collection name.